feat: implement delivery, order, payment services with BigDecimal and error handling
- Add Delivery service with BigDecimal for cost calculation
- Add Order service with Feign clients integration
- Add Payment service with BigDecimal support
- Fix code review comments:
  - Replace Double with BigDecimal for monetary values
  - Add ErrorDto for exception handling
  - Extract magic strings to constants
  - Add JavaDoc for DTOs
  - Move @ExceptionHandler to @ControllerAdvice
- Update Gateway routes for new services
- Update Docker Compose configuration

+++ commerce/delivery/src/main/java/ru/yandex/practicum/delivery/controller/DeliveryController.java
@@ -8,6 +8,7 @@ import ru.yandex.practicum.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.delivery.service.DeliveryService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @RestController
@@ -26,7 +27,7 @@ public class DeliveryController implements DeliveryClient {
 
     @Override
     @PostMapping("/cost")
-    public Double deliveryCost(@RequestBody OrderDto order) {
+    public BigDecimal deliveryCost(@RequestBody OrderDto order) {
         log.info("POST /api/v1/delivery/cost: orderId={}", order.getOrderId());
         return deliveryService.deliveryCost(order);
     }
