+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/payment/PaymentDto.java
@@ -5,16 +5,41 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
+/**
+ * Представление платежа в системе.
+ */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class PaymentDto {
+
+    /**
+     * Уникальный идентификатор платежа.
+     */
     private UUID paymentId;
-    private Double totalPayment;
-    private Double deliveryTotal;
-    private Double feeTotal;
-    private String state;  // PaymentState enum
+
+    /**
+     * Общая сумма к оплате (включая товары, доставку и налоги).
+     */
+    private BigDecimal totalPayment;
+
+    /**
+     * Стоимость доставки.
+     */
+    private BigDecimal deliveryTotal;
+
+    /**
+     * Сумма налогов и сборов.
+     */
+    private BigDecimal feeTotal;
+
+    /**
+     * Статус платежа.
+     * Возможные значения: PENDING, SUCCESS, FAILED
+     */
+    private String state;
 }
