+++ commerce/order/src/main/java/ru/yandex/practicum/order/client/PaymentOrderFeignClient.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
+
 @FeignClient(name = "payment")
 public interface PaymentOrderFeignClient {
 
@@ -13,8 +15,8 @@ public interface PaymentOrderFeignClient {
     PaymentDto paying(@RequestBody OrderDto order);
 
     @PostMapping("/api/v1/paying/totalCost")
-    Double getTotalCost(@RequestBody OrderDto order);
+    BigDecimal getTotalCost(@RequestBody OrderDto order);
 
     @PostMapping("/api/v1/pay/productCost")
-    Double productCost(@RequestBody OrderDto order);
+    BigDecimal productCost(@RequestBody OrderDto order);
 }
